pyproject.toml
[project]
name = "job-copilot-backend"
version = "0.1.0"
description = "AI-Powered Universal Job Application Copilot Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "gunicorn",
    "pydantic[email]",
    "pydantic-settings",
    "langchain",
    "langgraph",
    "langchain-openai",
    "langchain-anthropic",
    "openai",
    "anthropic",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "pgvector",
    "redis",
    "hiredis",
    "python-dotenv",
    "python-multipart",
    "httpx",
    "tenacity",
    "prometheus-client",
    "python-json-logger",
    "sentry-sdk[fastapi]",
    "python-jose[cryptography]",
    "passlib[argon2]",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
    "aiosqlite",
    "greenlet",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "agents", "db", "embeddings", "utils", "alembic"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (require database)",
    "unit: marks tests as unit tests (no external dependencies)",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
